+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * [#201](https://github.com/xenit-eu/alfresco-docker-gradle-plugin/issues/201) - Make it possible to disable rewriting log4j.properties
 
+### Fixed
+
+* [#229](https://github.com/xenit-eu/alfresco-docker-gradle-plugin/issues/229) - Overwriting files in WAR with AMP broken since v5.3.0
+
 ## Version 5.3.0 - 2021-04-21
 
 **This release drops support for Gradle versions before 5.6**